This commit contains Exception Handling, Comments and Project Restructuring towards Clean Architecture
