+++ Sync/Application/DTO/AbbreviatedContactDTO.cs
