+++ Sync/Application/DTO/ContactQueryRequest.cs
