+++ Sync/Application/Features/Contact/Conditions/IsStateCondition.cs
