+++ Sync/Core/Entities/AbbreviatedContact.cs
