+++ Sync/Core/ValueObjects/Group.cs
