+++ Sync/Infrastructure/Configuration.cs
