--- Sync/VirtuousService.cs
@@ -1,58 +1,0 @@
-﻿using AutoMapper;
-using RestSharp;
-using Sync.DTO;
-using System.Collections.Generic;
-using System.Diagnostics.Contracts;
-using System.Threading.Tasks;
-
-namespace Sync
-{
-    /// <summary>
-    /// API Docs found at https://docs.virtuoussoftware.com/
-    /// </summary>
-    internal class VirtuousService
-    {
-        private readonly RestClient _restClient;
-
-        public VirtuousService(IConfiguration configuration)
-        {
-            var apiBaseUrl = configuration.GetValue("VirtuousApiBaseUrl");
-            var apiKey = configuration.GetValue("VirtuousApiKey");
-
-            var options = new RestClientOptions(apiBaseUrl)
-            {
-                Authenticator = new RestSharp.Authenticators.OAuth2.OAuth2AuthorizationRequestHeaderAuthenticator(apiKey)
-            };
-
-            _restClient = new RestClient(options);
-        }
-
-        public async Task<PagedResult<AbbreviatedContactDTO>> GetContactsAsync(int skip, int take)
-        {
-            var request = new RestRequest("/api/Contact/Query", Method.Post);
-            request.AddQueryParameter("Skip", skip);
-            request.AddQueryParameter("Take", take);
-
-            var body = new ContactQueryRequest();
-            request.AddJsonBody(body);
-
-            var response = await _restClient.PostAsync<PagedResult<AbbreviatedContactDTO>>(request);
-            return response;
-        }
-        public async Task<PagedResult<AbbreviatedContactDTO>> GetContactsAsync(int skip, int take, List<Group> queryGroups)
-        {
-            if (queryGroups == null || queryGroups.Count == 0) throw new System.ArgumentNullException(); // TODO: create custom or more explicit exceptions.
-
-            var request = new RestRequest("/api/Contact/Query", Method.Post);
-            request.AddQueryParameter("Skip", skip);
-            request.AddQueryParameter("Take", take);
-
-            var body = new ContactQueryRequest();
-            body.Groups.AddRange(queryGroups);
-            request.AddJsonBody(body);
-
-            var response = await _restClient.PostAsync<PagedResult<AbbreviatedContactDTO>>(request);
-            return response;
-        }
-    }
-}
