+++ Sync/packages.config
@@ -3,9 +3,11 @@
   <package id="AutoMapper" version="10.1.1" targetFramework="net48" />
   <package id="CsvHelper" version="30.0.1" targetFramework="net48" />
   <package id="EntityFramework" version="6.5.1" targetFramework="net48" />
-  <package id="Microsoft.Bcl.AsyncInterfaces" version="7.0.0" targetFramework="net48" />
+  <package id="Microsoft.Bcl.AsyncInterfaces" version="9.0.1" targetFramework="net48" />
   <package id="Microsoft.Bcl.HashCode" version="1.0.0" targetFramework="net48" />
   <package id="Microsoft.CSharp" version="4.3.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.DependencyInjection" version="9.0.1" targetFramework="net48" />
+  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="9.0.1" targetFramework="net48" />
   <package id="RestSharp" version="110.2.0" targetFramework="net48" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net48" />
   <package id="System.Memory" version="4.5.5" targetFramework="net48" />
